+++ tests/test_datastore.py
@@ -563,7 +563,7 @@ def test_resample_datastore():
         directory=filepath, timezone_netcdf='UTC', file_ext='*.xml')
     assert ds.time.size == 3
 
-    ds_resampled = DataStore(xr.Dataset(ds).resample(time="47S").mean())
+    ds_resampled = DataStore(ds.resample(time="47S").mean())
 
     assert ds_resampled.time.size == 2
     assert ds_resampled.st.dims == ('x', 'time'), 'The dimension have to ' \